pyproject.toml
[project]
name = "neironggongchang"
version = "0.1.0"
description = "ShortVideo Studio - 数字人口播一页式工作流"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "streamlit>=1.36",
    "openai>=1.40",
    "yt-dlp>=2025.1.0",
    "ffmpeg-python>=0.2",
    "requests>=2.32",
    "pytest>=9.0",
]
